requirements.txt
Django == 2.2

mysqlclient == 1.4.6

celery == 4.3.0

redis == 3.3.11

django_environ == 0.4.5

djangorestframework == 3.9.0

gunicorn == 19.9.0
